refactored working filter
