+++ mobile-app/app/screens/AppNavigator/screens/Balances/BalancesScreen.tsx
@@ -1,16 +1,14 @@
-import { getNativeIcon } from '@components/icons/assets'
+
 import { View } from '@components'
 import {
   ThemedScrollView,
-  ThemedSectionTitle,
-  ThemedTouchableOpacity
+  ThemedSectionTitle
 } from '@components/themed'
 import { useDisplayBalancesContext } from '@contexts/DisplayBalancesContext'
 import { useWalletContext } from '@shared-contexts/WalletContext'
 import { useWalletPersistenceContext } from '@shared-contexts/WalletPersistenceContext'
 import { useWhaleApiClient } from '@shared-contexts/WhaleContext'
 import { useBottomTabBarHeight } from '@react-navigation/bottom-tabs'
-import { StackNavigationProp, StackScreenProps } from '@react-navigation/stack'
 import { ocean } from '@store/ocean'
 import { fetchTokens, tokensSelector, WalletToken } from '@store/wallet'
 import { tailwind } from '@tailwind'
@@ -23,13 +21,13 @@ import { DFIBalanceCard } from '@screens/AppNavigator/screens/Balances/component
 import { translate } from '@translations'
 import { RefreshControl } from 'react-native'
 import { BalanceControlCard } from '@screens/AppNavigator/screens/Balances/components/BalanceControlCard'
-import { EmptyBalances } from '@screens/AppNavigator/screens/Balances/components/EmptyBalances'
+
 import { RootState } from '@store'
 import { useTokenPrice } from './hooks/TokenPrice'
-import { TokenNameText } from '@screens/AppNavigator/screens/Balances/components/TokenNameText'
-import { TokenAmountText } from '@screens/AppNavigator/screens/Balances/components/TokenAmountText'
+
+import { StackScreenProps } from '@react-navigation/stack'
+
 import { TotalPortfolio } from './components/TotalPortfolio'
-import { SkeletonLoader, SkeletonLoaderScreen } from '@components/SkeletonLoader'
 import { fetchVaults, LoanVault, vaultsSelector } from '@store/loans'
 import { LoanVaultState, LoanVaultTokenAmount } from '@defichain/whale-api-client/dist/api/loan'
 import { useIsFocused } from '@react-navigation/native'
@@ -188,10 +186,9 @@ export function BalancesScreen ({ navigation }: Props): JSX.Element {
       />
       <ThemedSectionTitle text={translate('screens/BalancesScreen', 'YOUR ASSETS')} style={tailwind('px-4 pt-2 pb-2 text-xs font-medium')} />
       <DFIBalanceCard />
-      {/* filter tab  */}
       <View style={tailwind('flex')}>
         <BalanceCards
-          dstTokens={filteredTokens}
+          filteredTokens={filteredTokens}
           navigation={navigation}
           buttonGroupOptions={{
             activeButtonGroup: activeButtonGroup,
@@ -199,81 +196,8 @@ export function BalancesScreen ({ navigation }: Props): JSX.Element {
             onButtonGroupPress: handleButtonFilter
           }}
         />
-        <BalanceList dstTokens={filteredTokens} navigation={navigation} />
       </View>
 
     </ThemedScrollView>
-  )
-}
-
-// TODO: refactoring this function into BalanceCards component
-function BalanceList ({
-  dstTokens,
-  navigation
-}: { dstTokens: BalanceRowToken[], navigation: StackNavigationProp<BalanceParamList> }): JSX.Element {
-  const { hasFetchedToken } = useSelector((state: RootState) => (state.wallet))
-
-  if (!hasFetchedToken) {
-    return (
-      <View style={tailwind('px-4 py-1.5 -mb-3')}>
-        <SkeletonLoader row={4} screen={SkeletonLoaderScreen.Balance} />
-      </View>
-    )
-  }
-
-  return (
-    <>
-      {
-        dstTokens.length === 0
-          ? (
-            <EmptyBalances />
-          )
-          : (
-            <>
-              <View testID='card_balance_row_container'>
-                {dstTokens.sort((a, b) => new BigNumber(b.usdAmount).minus(new BigNumber(a.usdAmount)).toNumber()).map((item) => (
-                  <View key={item.symbol} style={tailwind('p-4 pt-1.5 pb-1.5')}>
-                    <BalanceItemRow
-                      onPress={() => navigation.navigate({
-                        name: 'TokenDetail',
-                        params: { token: item },
-                        merge: true
-                      })}
-                      token={item}
-                    />
-                  </View>
-                ))}
-              </View>
-            </>
-            )
-      }
-    </>
-  )
-}
-
-function BalanceItemRow ({
-  token,
-  onPress
-}: { token: BalanceRowToken, onPress: () => void }): JSX.Element {
-  const Icon = getNativeIcon(token.displaySymbol)
-  const testID = `balances_row_${token.id}`
-  const { isBalancesDisplayed } = useDisplayBalancesContext()
-  return (
-    <ThemedTouchableOpacity
-      dark={tailwind('bg-gray-800')}
-      light={tailwind('bg-white')}
-      onPress={onPress}
-      style={tailwind('p-4 rounded-lg flex-row justify-between items-center')}
-      testID={testID}
-    >
-      <View style={tailwind('flex-row items-center flex-grow')}>
-        <Icon testID={`${testID}_icon`} />
-        <TokenNameText displaySymbol={token.displaySymbol} name={token.name} testID={testID} />
-        <TokenAmountText
-          tokenAmount={token.amount} usdAmount={token.usdAmount} testID={testID}
-          isBalancesDisplayed={isBalancesDisplayed}
-        />
-      </View>
-    </ThemedTouchableOpacity>
   )
 }
