+++ mobile-app/app/screens/AppNavigator/screens/Balances/components/BalanceCards.tsx
@@ -1,22 +1,19 @@
-import { ThemedView } from '@components/themed'
+import { ThemedFlatList, ThemedTouchableOpacity } from '@components/themed'
 import { BalanceParamList } from '../BalancesNavigator'
 import { BalanceRowToken } from '../BalancesScreen'
 import { StackNavigationProp } from '@react-navigation/stack'
-
-// import { EmptyBalances } from './EmptyBalances'
-// import BigNumber from 'bignumber.js'
-
-// import { RootState } from '@store'
-// import { useSelector } from 'react-redux'
-// import { useState } from 'react'
-// import { SkeletonLoader, SkeletonLoaderScreen } from '@components/SkeletonLoader'
-
 import { View } from '@components'
-
 import { translate } from '@translations'
 import tailwind from 'tailwind-rn'
 import { ButtonGroup } from '../../Dex/components/ButtonGroup'
-// import { Platform } from 'react-native'
+import { RootState } from '@store'
+import { useSelector } from 'react-redux'
+import { SkeletonLoader, SkeletonLoaderScreen } from '@components/SkeletonLoader'
+import { EmptyBalances } from './EmptyBalances'
+import { TokenNameText } from '@screens/AppNavigator/screens/Balances/components/TokenNameText'
+import { TokenAmountText } from '@screens/AppNavigator/screens/Balances/components/TokenAmountText'
+import { useDisplayBalancesContext } from '@contexts/DisplayBalancesContext'
+import { getNativeIcon } from '@components/icons/assets'
 
 export enum ButtonGroupTabKey {
   AllTokens = 'ALL_TOKENS',
@@ -26,11 +23,8 @@ export enum ButtonGroupTabKey {
 }
 
 interface BalanceCardProps {
-  // passed props into BalanceList from Balance screen
-  dstTokens: BalanceRowToken[]
+  filteredTokens: BalanceRowToken[]
   navigation: StackNavigationProp<BalanceParamList>
-
-  // required props for tabs - need to pass props from balance screen too
   buttonGroupOptions?: {
     onButtonGroupPress: (key: ButtonGroupTabKey) => void
     activeButtonGroup: string
@@ -39,11 +33,8 @@ interface BalanceCardProps {
 }
 
 export function BalanceCards ({
-  // passed props into BalanceList from Balance screen
-  dstTokens,
+  filteredTokens,
   navigation,
-
-  // required props for tabs - need to pass props from balance screen too
   buttonGroupOptions
 }: BalanceCardProps): JSX.Element {
   const buttonGroup = [
@@ -75,22 +66,103 @@ export function BalanceCards ({
     }
   }
 
-  return (
-    <ThemedView
-      light={tailwind('bg-gray-50')}
-      style={tailwind('p-4')}
+  const { hasFetchedToken } = useSelector((state: RootState) => (state.wallet))
 
-    >{
-        buttonGroupOptions !== undefined &&
-          <View style={tailwind('text-xs text-center rounded-2xl')}>
-            <ButtonGroup
-              buttons={buttonGroup}
-              activeButtonGroupItem={buttonGroupOptions.activeButtonGroup}
-              modalStyle={tailwind('font-semibold text-xs py-1')} // style for smaller text
-              testID='balance_button_group'
-            />
-          </View>
+  const renderItem = ({
+    item,
+    index
+  }: {
+    item: BalanceRowToken
+    index: number
+  }): JSX.Element => {
+    console.log('item', item)
+    console.log('filteredTokens', filteredTokens)
+    console.log('=====')
+
+    return (
+
+      <View testID='card_balance_row_container'>
+
+        <View key={index}>
+          <BalanceItemRow
+            onPress={() => navigation.navigate({
+              name: 'TokenDetail',
+              params: { token: item },
+              merge: true
+            })}
+            token={item}
+          />
+        </View>
+      </View>
+    )
+  }
+
+  return (
+    <ThemedFlatList
+      light={tailwind('bg-gray-50')}
+      dark={tailwind('bg-gray-900')}
+      contentContainerStyle={tailwind('p-4 pb-2')}
+      data={filteredTokens}
+      numColumns={1}
+      windowSize={2}
+      // initialNumToRender={5}
+      // keyExtractor={(item) => item.data.id}
+      testID=''
+      renderItem={renderItem}
+      ListHeaderComponent={
+        <>
+          {
+            buttonGroupOptions !== undefined &&
+            (
+              <>
+                <View style={tailwind('mb-4')}>
+                  <ButtonGroup buttons={buttonGroup} activeButtonGroupItem={buttonGroupOptions.activeButtonGroup} testID='dex_button_group' />
+                </View>
+              </>
+            )
+}
+          {
+            !hasFetchedToken &&
+              <View style={tailwind('')}>
+                <SkeletonLoader row={4} screen={SkeletonLoaderScreen.Balance} />
+              </View>
+          }
+          {
+            // display empty balance
+            filteredTokens.length === 0 &&
+              <EmptyBalances />
+
+          }
+
+        </>
       }
-    </ThemedView>
+    />
+  )
+}
+
+function BalanceItemRow ({
+  token,
+  onPress
+}: { token: BalanceRowToken, onPress: () => void }): JSX.Element {
+  const Icon = getNativeIcon(token.displaySymbol)
+  const testID = `balances_row_${token.id}`
+  const { isBalancesDisplayed } = useDisplayBalancesContext()
+  return (
+    <ThemedTouchableOpacity
+      dark={tailwind('bg-gray-800')}
+      light={tailwind('bg-white')}
+      onPress={onPress}
+      style={tailwind('p-4 rounded-lg flex-row justify-between items-center')}
+      testID={testID}
+    >
+      <View style={tailwind('flex-row items-center flex-grow')}>
+        <Icon testID={`${testID}_icon`} />
+        <TokenNameText displaySymbol={token.displaySymbol} name={token.name} testID={testID} />
+        <TokenAmountText
+          tokenAmount={token.amount} usdAmount={token.usdAmount} testID={testID}
+          isBalancesDisplayed={isBalancesDisplayed}
+        />
+      </View>
+    </ThemedTouchableOpacity>
   )
 }
